+++ functional_widget/example/lib/main.g.dart
@@ -14,11 +14,6 @@ class Foo extends StatelessWidget {
   @override
   Widget build(BuildContext _context) => foo(value);
   @override
-  int get hashCode => value.hashCode;
-  @override
-  bool operator ==(Object o) =>
-      identical(o, this) || (o is Foo && value == o.value);
-  @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties.add(IntProperty('value', value));
@@ -38,15 +33,6 @@ class Example extends StatelessWidget {
   @override
   Widget build(BuildContext _context) =>
       example(foo, bar, onChanged: onChanged);
-  @override
-  int get hashCode => hashValues(foo, bar, onChanged);
-  @override
-  bool operator ==(Object o) =>
-      identical(o, this) ||
-      (o is Example &&
-          foo == o.foo &&
-          bar == o.bar &&
-          onChanged == o.onChanged);
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);