+++ functional_widget/lib/src/utils.dart
@@ -17,11 +17,9 @@ FunctionalWidget parseBuilderOptions(BuilderOptions options) {
   final widgetType = _parseWidgetType(options.config['widgetType']);
   final debugFillProperties =
       _parseDebugFillProperties(options.config['debugFillProperties']);
-  final equality = _parseEquality(options.config['equality']);
   return FunctionalWidget(
     widgetType: widgetType,
     debugFillProperties: debugFillProperties,
-    equality: equality,
   );
 }
 
@@ -35,24 +33,6 @@ bool _parseDebugFillProperties(dynamic value) {
   }
   throw ArgumentError.value(value, 'debugFillProperties',
       'Invalid value. Potential values are `true` or `false`');
-}
-
-FunctionalWidgetEquality _parseEquality(dynamic value) {
-  if (value == null) {
-    return null;
-  }
-  if (value is String) {
-    switch (value) {
-      case 'none':
-        return FunctionalWidgetEquality.none;
-      case 'equal':
-        return FunctionalWidgetEquality.equal;
-      case 'identical':
-        return FunctionalWidgetEquality.identical;
-    }
-  }
-  throw ArgumentError.value(value, 'widgetType',
-      'Invalid value. Potential values are `none`/`equal`/`identical`');
 }
 
 FunctionalWidgetType _parseWidgetType(dynamic value) {
@@ -75,8 +55,6 @@ FunctionalWidget parseFunctionalWidetAnnotation(ConstantReader reader) {
   return FunctionalWidget(
     widgetType:
         _parseEnum(reader.read('widgetType'), FunctionalWidgetType.values),
-    equality:
-        _parseEnum(reader.read('equality'), FunctionalWidgetEquality.values),
   );
 }